+++ NPCMapLocations/docs/release-notes.md
@@ -8,6 +8,7 @@
 * Fixed minimap no longer becoming transparent when you hover the cursor over it.
 * Fixed raccoon icon on the map (thanks to Exblosis!).
 * Fixed invisible NPCs shown on the map (thanks to Abagaianye!).
+* Fixed 'host does not have NPC Map Locations installed' check on the host.
 * Improved translations. Thanks to burunduk (updated Ukrainian), CaranudLapin (updated French), CuongNguyen0903 (added Vietnamese), DARKS0UND (updated Korean), Exblosis (updated German), juliowh (updated Portuguese), MagiaBaiser (updated Spanish), mitekano23 (updated Japanese), and ykinsoul (updated Chinese!).
 
 ## 3.1.0